changed to sentence case after rereading the ticket
